projecteuler.15

+++ src/projecteuler/p15/p15.cpp
@@ -1,8 +1,8 @@
 #include <iostream>
 int main(){
-  long answer = 1;
+  long long answer = 1;
   for(int i = 1; i < 21; ++i) {
-    answer = answer * (41 - i) /  i;
+    answer = answer * (41 - i) / i;
   }
   std::cout << answer << std::endl;
   return 0;
